move NormalScanConfig component to folder
