+++ frontend/components/MicroscopeMapDisplay.jsx
@@ -5,6 +5,7 @@ import ArtifactZarrLoader from '../services/artifactZarrLoader.js';
 import LayerPanel from './microscope/map/LayerPanel';
 import useExperimentZarrManager from './microscope/map/ExperimentZarrManager';
 import QuickScanConfig from './microscope/controls/QuickScanConfig';
+import NormalScanConfig from './microscope/controls/NormalScanConfig';
 import './MicroscopeMapDisplay.css';
 
 const MicroscopeMapDisplay = ({
@@ -2137,60 +2138,12 @@ const MicroscopeMapDisplay = ({
   const [gridDragStart, setGridDragStart] = useState(null);
   const [gridDragEnd, setGridDragEnd] = useState(null);
   const [isGridDragging, setIsGridDragging] = useState(false);
+  const [gridSelectedCells, setGridSelectedCells] = useState({});
 
-  // Compute selected cells for grid drag
-  const gridSelectedCells = useMemo(() => {
-    if (!gridDragStart || !gridDragEnd) return {};
-    const layout = getWellPlateLayout();
-    const r1 = Math.min(gridDragStart[0], gridDragEnd[0]);
-    const c1 = Math.min(gridDragStart[1], gridDragEnd[1]);
-    const r2 = Math.max(gridDragStart[0], gridDragEnd[0]);
-    const c2 = Math.max(gridDragStart[1], gridDragEnd[1]);
-    const selected = {};
-    for (let r = r1; r <= r2; r++) {
-      for (let c = c1; c <= c2; c++) {
-        selected[`${r}-${c}`] = true;
-      }
-    }
-    return selected;
-  }, [gridDragStart, gridDragEnd, getWellPlateLayout]);
+  // Grid selected cells are managed by state in NormalScanConfig component
 
   // Mouse handlers for grid selection
-  const handleGridCellMouseDown = (rowIdx, colIdx) => {
-    setGridDragStart([rowIdx, colIdx]);
-    setGridDragEnd([rowIdx, colIdx]);
-    setIsGridDragging(true);
-  };
-  const handleGridCellMouseEnter = (rowIdx, colIdx) => {
-    if (isGridDragging) {
-      setGridDragEnd([rowIdx, colIdx]);
-    }
-  };
-  useEffect(() => {
-    const handleMouseUp = () => {
-      if (isGridDragging && gridDragStart && gridDragEnd) {
-        // Compute all selected wells
-        const layout = getWellPlateLayout();
-        const r1 = Math.min(gridDragStart[0], gridDragEnd[0]);
-        const c1 = Math.min(gridDragStart[1], gridDragEnd[1]);
-        const r2 = Math.max(gridDragStart[0], gridDragEnd[0]);
-        const c2 = Math.max(gridDragStart[1], gridDragEnd[1]);
-        const newSelected = [];
-        for (let r = r1; r <= r2; r++) {
-          for (let c = c1; c <= c2; c++) {
-            newSelected.push(getWellIdFromIndex(r, c));
-          }
-        }
-        // Add to previous selection (union)
-        setSelectedWells(prev => Array.from(new Set([...prev, ...newSelected])));
-      }
-      setIsGridDragging(false);
-      setGridDragStart(null);
-      setGridDragEnd(null);
-    };
-    window.addEventListener('mouseup', handleMouseUp);
-    return () => window.removeEventListener('mouseup', handleMouseUp);
-  }, [isGridDragging, gridDragStart, gridDragEnd, getWellIdFromIndex, getWellPlateLayout]);
+  // Grid selection handlers are handled in NormalScanConfig component
 
   // In scan start button handler, pass selectedWells to normal_scan_with_stitching
 
@@ -4152,607 +4105,63 @@ const MicroscopeMapDisplay = ({
         quickFpsInput={quickFpsInput}
       />
 
-      {/* Scan Configuration Side Panel */}
-      {showScanConfig && (
-        <div className="absolute top-12 right-2 bg-gray-800 border border-gray-600 rounded-lg shadow-xl w-80 z-50 text-white scan-config-panel">
-          <div className="flex items-center justify-between p-3 border-b border-gray-600 flex-shrink-0">
-            <h3 className="text-sm font-semibold text-gray-200">Scan Configuration</h3>
-            <button
-              onClick={() => {
-                setShowScanConfig(false);
-                setIsRectangleSelection(false);
-                setRectangleStart(null);
-                setRectangleEnd(null);
-                setDragSelectedWell(null);
-                // Clean up grid drawing states
-                setGridDragStart(null);
-                setGridDragEnd(null);
-                setIsGridDragging(false);
-              }}
-              className="text-gray-400 hover:text-white p-1"
-              title="Close"
-            >
-              <i className="fas fa-times"></i>
-            </button>
-          </div>
-          
-          <div className="p-3 scan-config-content">
-            {/* --- Multi-well Selection Grid UI --- */}
-            <div className="scan-well-plate-grid-container">
-              <div className="flex flex-col w-full items-center">
-                <div className="flex w-full items-center mb-1">
-                  <span className="text-xs text-gray-300 mr-2">Selected: {selectedWells.length}</span>
-                  <button
-                    onClick={() => setSelectedWells([])}
-                    className="ml-auto px-2 py-1 text-xs bg-gray-600 hover:bg-gray-500 text-white rounded disabled:opacity-50 disabled:cursor-not-allowed"
-                    title="Clear all well selections"
-                    disabled={selectedWells.length === 0}
-                  >
-                    <i className="fas fa-refresh mr-1"></i>Refresh
-                  </button>
-                </div>
-                <div className="scan-well-plate-grid">
-                  <div className="scan-grid-col-labels">
-                    <div></div>
-                    {getWellPlateGridLabels().cols.map((label, colIdx) => (
-                      <div key={`col-${label}`} className="scan-grid-label">{label}</div>
-                    ))}
-                  </div>
-                  {getWellPlateGridLabels().rows.map((rowLabel, rowIdx) => (
-                    <div key={`row-${rowIdx}`} className="scan-grid-row">
-                      <div className="scan-grid-label">{rowLabel}</div>
-                      {getWellPlateGridLabels().cols.map((colLabel, colIdx) => {
-                        const wellId = getWellIdFromIndex(rowIdx, colIdx);
-                        const isSelected = selectedWells.includes(wellId);
-                        const isDragSelected = gridSelectedCells[`${rowIdx}-${colIdx}`];
-                        return (
-                          <div
-                            key={`cell-${rowIdx}-${colIdx}`}
-                            className={`scan-grid-cell${isSelected || isDragSelected ? ' selected' : ''}`}
-                            onMouseDown={() => handleGridCellMouseDown(rowIdx, colIdx)}
-                            onMouseEnter={() => handleGridCellMouseEnter(rowIdx, colIdx)}
-                            style={{ userSelect: 'none' }}
-                          >
-                            {/* Optionally show wellId or leave blank for cleaner look */}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
+      {/* Normal Scan Configuration Side Panel */}
+      <NormalScanConfig
+        // State props
+        showScanConfig={showScanConfig}
+        setShowScanConfig={setShowScanConfig}
+        scanParameters={scanParameters}
+        setScanParameters={setScanParameters}
+        isScanInProgress={isScanInProgress}
+        setIsScanInProgress={setIsScanInProgress}
+        activeExperiment={activeExperiment}
+        wellPaddingMm={wellPaddingMm}
+        wellPlateType={wellPlateType}
+        selectedWells={selectedWells}
+        setSelectedWells={setSelectedWells}
+        isRectangleSelection={isRectangleSelection}
+        setIsRectangleSelection={setIsRectangleSelection}
+        rectangleStart={rectangleStart}
+        setRectangleStart={setRectangleStart}
+        rectangleEnd={rectangleEnd}
+        setRectangleEnd={setRectangleEnd}
+        dragSelectedWell={dragSelectedWell}
+        setDragSelectedWell={setDragSelectedWell}
+        gridSelectedCells={gridSelectedCells}
+        setGridSelectedCells={setGridSelectedCells}
+        gridDragStart={gridDragStart}
+        setGridDragStart={setGridDragStart}
+        gridDragEnd={gridDragEnd}
+        setGridDragEnd={setGridDragEnd}
+        isGridDragging={isGridDragging}
+        setIsGridDragging={setIsGridDragging}
+        visibleLayers={visibleLayers}
+        setVisibleLayers={setVisibleLayers}
+        refreshScanResults={refreshScanResults}
+        
+        // Service props
+        microscopeControlService={microscopeControlService}
+        appendLog={appendLog}
+        showNotification={showNotification}
+        isWebRtcActive={isWebRtcActive}
+        toggleWebRtcStream={toggleWebRtcStream}
+        setMicroscopeBusy={setMicroscopeBusy}
+        setCurrentOperation={setCurrentOperation}
+        
+        // Input validation hooks
+        startXInput={startXInput}
+        startYInput={startYInput}
+        nxInput={nxInput}
+        nyInput={nyInput}
+        dxInput={dxInput}
+        dyInput={dyInput}
+        
+        // Helper functions
+        getWellPlateGridLabels={getWellPlateGridLabels}
+        getWellIdFromIndex={getWellIdFromIndex}
+        loadCurrentMicroscopeSettings={loadCurrentMicroscopeSettings}
+      />
 
-            <div className="space-y-2 text-xs">
-              <div>
-                <label className="block text-gray-300 font-medium mb-1">Start Position (mm)</label>
-                <div className="flex space-x-2">
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={startXInput.inputValue}
-                      onChange={startXInput.handleInputChange}
-                      onKeyDown={startXInput.handleKeyDown}
-                      onBlur={startXInput.handleBlur}
-                      className={getInputValidationClasses(
-                        startXInput.isValid,
-                        startXInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      step="0.1"
-                      disabled={isScanInProgress}
-                      placeholder="X position"
-                    />
-                  </div>
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={startYInput.inputValue}
-                      onChange={startYInput.handleInputChange}
-                      onKeyDown={startYInput.handleKeyDown}
-                      onBlur={startYInput.handleBlur}
-                      className={getInputValidationClasses(
-                        startYInput.isValid,
-                        startYInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      step="0.1"
-                      disabled={isScanInProgress}
-                      placeholder="Y position"
-                    />
-                  </div>
-                </div>
-              </div>
-              
-              <div>
-                <label className="block text-gray-300 font-medium mb-1">Grid Size (positions)</label>
-                <div className="flex space-x-2">
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={nxInput.inputValue}
-                      onChange={nxInput.handleInputChange}
-                      onKeyDown={nxInput.handleKeyDown}
-                      onBlur={nxInput.handleBlur}
-                      className={getInputValidationClasses(
-                        nxInput.isValid,
-                        nxInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      min="1"
-                      disabled={isScanInProgress}
-                      placeholder="Nx"
-                    />
-                  </div>
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={nyInput.inputValue}
-                      onChange={nyInput.handleInputChange}
-                      onKeyDown={nyInput.handleKeyDown}
-                      onBlur={nyInput.handleBlur}
-                      className={getInputValidationClasses(
-                        nyInput.isValid,
-                        nyInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      min="1"
-                      disabled={isScanInProgress}
-                      placeholder="Ny"
-                    />
-                  </div>
-                </div>
-              </div>
-              
-              <div>
-                <label className="block text-gray-300 font-medium mb-1">Step Size (mm)</label>
-                <div className="flex space-x-2">
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={dxInput.inputValue}
-                      onChange={dxInput.handleInputChange}
-                      onKeyDown={dxInput.handleKeyDown}
-                      onBlur={dxInput.handleBlur}
-                      className={getInputValidationClasses(
-                        dxInput.isValid,
-                        dxInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      step="0.1"
-                      min="0.1"
-                      disabled={isScanInProgress}
-                      placeholder="dX step"
-                    />
-                  </div>
-                  <div className="w-1/2 input-validation-container">
-                    <input
-                      type="number"
-                      value={dyInput.inputValue}
-                      onChange={dyInput.handleInputChange}
-                      onKeyDown={dyInput.handleKeyDown}
-                      onBlur={dyInput.handleBlur}
-                      className={getInputValidationClasses(
-                        dyInput.isValid,
-                        dyInput.hasUnsavedChanges,
-                        "w-full px-2 py-1 bg-gray-700 border rounded text-white"
-                      )}
-                      step="0.1"
-                      min="0.1"
-                      disabled={isScanInProgress}
-                      placeholder="dY step"
-                    />
-                  </div>
-                </div>
-              </div>
-              
-
-              
-              <div>
-                <div className="flex items-center justify-between mb-2">
-                  <label className="block text-gray-300 font-medium">Illumination Channels</label>
-                  <div className="flex space-x-1">
-                    <button
-                      onClick={() => {
-                        if (isScanInProgress) return;
-                        loadCurrentMicroscopeSettings();
-                      }}
-                      className="px-2 py-1 text-xs bg-green-600 hover:bg-green-500 text-white rounded disabled:opacity-50 disabled:cursor-not-allowed"
-                      disabled={isScanInProgress || !microscopeControlService}
-                      title="Load current microscope settings"
-                    >
-                      <i className="fas fa-download mr-1"></i>
-                      Load Current
-                    </button>
-                    <button
-                      onClick={() => {
-                        if (isScanInProgress) return;
-                        // Find a channel that's not already in use
-                        const availableChannels = [
-                          'BF LED matrix full',
-                          'Fluorescence 405 nm Ex',
-                          'Fluorescence 488 nm Ex',
-                          'Fluorescence 561 nm Ex',
-                          'Fluorescence 638 nm Ex',
-                          'Fluorescence 730 nm Ex'
-                        ];
-                        const usedChannels = scanParameters.illumination_settings.map(s => s.channel);
-                        const nextChannel = availableChannels.find(c => !usedChannels.includes(c)) || 'BF LED matrix full';
-                        
-                        setScanParameters(prev => ({
-                          ...prev,
-                          illumination_settings: [
-                            ...prev.illumination_settings,
-                            {
-                              channel: nextChannel,
-                              intensity: 50,
-                              exposure_time: 100
-                            }
-                          ]
-                        }));
-                      }}
-                      className="px-2 py-1 text-xs bg-blue-600 hover:bg-blue-500 text-white rounded disabled:opacity-50 disabled:cursor-not-allowed"
-                      disabled={isScanInProgress || scanParameters.illumination_settings.length >= 6}
-                      title="Add channel"
-                    >
-                      <i className="fas fa-plus mr-1"></i>
-                      Add Channel
-                    </button>
-                  </div>
-                </div>
-                
-                <div className="space-y-2 max-h-48 overflow-y-auto">
-                  {scanParameters.illumination_settings.map((setting, index) => (
-                    <div key={index} className="bg-gray-700 p-2 rounded border border-gray-600">
-                      <div className="flex items-center justify-between mb-2">
-                        <span className="text-xs text-gray-300 font-medium">Channel {index + 1}</span>
-                        <button
-                          onClick={() => {
-                            if (isScanInProgress || scanParameters.illumination_settings.length <= 1) return;
-                            setScanParameters(prev => ({
-                              ...prev,
-                              illumination_settings: prev.illumination_settings.filter((_, i) => i !== index)
-                            }));
-                          }}
-                          className="px-1 py-0.5 text-xs bg-red-600 hover:bg-red-500 text-white rounded disabled:opacity-50 disabled:cursor-not-allowed"
-                          disabled={isScanInProgress || scanParameters.illumination_settings.length <= 1}
-                          title="Remove channel"
-                        >
-                          <i className="fas fa-times"></i>
-                        </button>
-                      </div>
-                      
-                      <div className="space-y-2">
-                        <div className="relative">
-                          <select
-                            value={setting.channel}
-                            onChange={(e) => {
-                              if (isScanInProgress) return;
-                              setScanParameters(prev => ({
-                                ...prev,
-                                illumination_settings: prev.illumination_settings.map((s, i) => 
-                                  i === index ? { ...s, channel: e.target.value } : s
-                                )
-                              }));
-                            }}
-                            className={`w-full px-2 py-1 text-xs bg-gray-600 border rounded text-white ${
-                              scanParameters.illumination_settings.filter(s => s.channel === setting.channel).length > 1 
-                                ? 'border-yellow-500' 
-                                : 'border-gray-500'
-                            }`}
-                            disabled={isScanInProgress}
-                          >
-                            <option value="BF LED matrix full">BF LED matrix full</option>
-                            <option value="Fluorescence 405 nm Ex">Fluorescence 405 nm Ex</option>
-                            <option value="Fluorescence 488 nm Ex">Fluorescence 488 nm Ex</option>
-                            <option value="Fluorescence 561 nm Ex">Fluorescence 561 nm Ex</option>
-                            <option value="Fluorescence 638 nm Ex">Fluorescence 638 nm Ex</option>
-                            <option value="Fluorescence 730 nm Ex">Fluorescence 730 nm Ex</option>
-                          </select>
-                          {scanParameters.illumination_settings.filter(s => s.channel === setting.channel).length > 1 && (
-                            <div className="absolute -top-1 -right-1">
-                              <i className="fas fa-exclamation-triangle text-yellow-500 text-xs" title="Duplicate channel detected"></i>
-                            </div>
-                          )}
-                        </div>
-                        
-                        <div className="flex space-x-2">
-                          <div className="flex-1">
-                            <input
-                              type="number"
-                              value={setting.intensity}
-                              onChange={(e) => {
-                                if (isScanInProgress) return;
-                                const value = parseInt(e.target.value) || 0;
-                                if (value >= 1 && value <= 100) {
-                                  setScanParameters(prev => ({
-                                    ...prev,
-                                    illumination_settings: prev.illumination_settings.map((s, i) => 
-                                      i === index ? { ...s, intensity: value } : s
-                                    )
-                                  }));
-                                }
-                              }}
-                              className="w-full px-2 py-1 text-xs bg-gray-600 border border-gray-500 rounded text-white"
-                              min="1"
-                              max="100"
-                              disabled={isScanInProgress}
-                              placeholder="Intensity %"
-                            />
-                          </div>
-                          <div className="flex-1">
-                            <input
-                              type="number"
-                              value={setting.exposure_time}
-                              onChange={(e) => {
-                                if (isScanInProgress) return;
-                                const value = parseInt(e.target.value) || 0;
-                                if (value >= 1 && value <= 1000) {
-                                  setScanParameters(prev => ({
-                                    ...prev,
-                                    illumination_settings: prev.illumination_settings.map((s, i) => 
-                                      i === index ? { ...s, exposure_time: value } : s
-                                    )
-                                  }));
-                                }
-                              }}
-                              className="w-full px-2 py-1 text-xs bg-gray-600 border border-gray-500 rounded text-white"
-                              min="1"
-                              max="1000"
-                              disabled={isScanInProgress}
-                              placeholder="Exposure ms"
-                            />
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-                
-                {scanParameters.illumination_settings.length > 1 && (
-                  <div className="mt-2 p-2 bg-blue-900 bg-opacity-30 rounded border border-blue-500 text-xs">
-                    <div className="flex items-center text-blue-300 mb-1">
-                      <i className="fas fa-info-circle mr-1"></i>
-                      Multi-channel Acquisition
-                    </div>
-                    <div className="text-gray-300">
-                      Channels: {scanParameters.illumination_settings.map(s => 
-                        s.channel.replace('Fluorescence ', '').replace(' Ex', '').replace('BF LED matrix full', 'BF')
-                      ).join(', ')}
-                    </div>
-                    {scanParameters.illumination_settings.some((setting, index) => 
-                      scanParameters.illumination_settings.filter(s => s.channel === setting.channel).length > 1
-                    ) && (
-                      <div className="text-yellow-300 mt-1">
-                        <i className="fas fa-exclamation-triangle mr-1"></i>
-                        Warning: Duplicate channels detected
-                      </div>
-                    )}
-                  </div>
-                )}
-              </div>
-              
-              <div className="flex items-center space-x-4">
-                <label className="flex items-center text-xs">
-                  <input
-                    type="checkbox"
-                    checked={scanParameters.do_contrast_autofocus}
-                    onChange={(e) => setScanParameters(prev => ({ ...prev, do_contrast_autofocus: e.target.checked }))}
-                    className="mr-2"
-                    disabled={isScanInProgress}
-                  />
-                  Contrast AF
-                </label>
-                <label className="flex items-center text-xs">
-                  <input
-                    type="checkbox"
-                    checked={scanParameters.do_reflection_af}
-                    onChange={(e) => setScanParameters(prev => ({ ...prev, do_reflection_af: e.target.checked }))}
-                    className="mr-2"
-                    disabled={isScanInProgress}
-                  />
-                  Reflection AF
-                </label>
-              </div>
-              
-              {/* Upload During Scanning Toggle */}
-              <div className="bg-gray-700 p-2 rounded">
-                <div className="text-purple-300 font-medium mb-1"><i className="fas fa-cloud-upload-alt mr-1"></i>Upload Settings</div>
-                <label className="flex items-center text-xs">
-                  <input
-                    type="checkbox"
-                    checked={scanParameters.uploading}
-                    onChange={(e) => setScanParameters(prev => ({ ...prev, uploading: e.target.checked }))}
-                    className="mr-2"
-                    disabled={isScanInProgress}
-                  />
-                  Upload during scanning
-                  <i className="fas fa-question-circle ml-1 text-gray-400" title="Enable background upload of scan data to artifact manager during scanning"></i>
-                </label>
-              </div>
-              
-              <div className="bg-gray-700 p-2 rounded text-xs">
-                <div>Total scan area: {(scanParameters.Nx * scanParameters.dx_mm).toFixed(1)} × {(scanParameters.Ny * scanParameters.dy_mm).toFixed(1)} mm</div>
-                <div>Total positions: {scanParameters.Nx * scanParameters.Ny}</div>
-                <div>Channels: {scanParameters.illumination_settings.length}</div>
-                <div>Total images: {scanParameters.Nx * scanParameters.Ny * scanParameters.illumination_settings.length}</div>
-                <div>End position: ({(scanParameters.start_x_mm + (scanParameters.Nx-1) * scanParameters.dx_mm).toFixed(1)}, {(scanParameters.start_y_mm + (scanParameters.Ny-1) * scanParameters.dy_mm).toFixed(1)}) mm</div>
-              </div>
-              
-              
-              {isRectangleSelection && (
-                <div className="bg-blue-900 bg-opacity-50 p-2 rounded text-xs border border-blue-500">
-                  <i className="fas fa-vector-square mr-1"></i>
-                  Drag on the map to select scan area. Current settings will be used as defaults.
-                </div>
-              )}
-            </div>
-            
-                        <div className="flex justify-end space-x-2 mt-4">
-                <button
-                  onClick={() => {
-                    if (isScanInProgress) return;
-                    if (isRectangleSelection) {
-                      // Stop rectangle selection
-                      setIsRectangleSelection(false);
-                      setRectangleStart(null);
-                      setRectangleEnd(null);
-                      setDragSelectedWell(null);
-                    } else {
-                      // Start rectangle selection - clear any existing selection first
-                      setRectangleStart(null);
-                      setRectangleEnd(null);
-                      setDragSelectedWell(null);
-                      setIsRectangleSelection(true);
-                    }
-                  }}
-                  className={`px-3 py-1 text-xs rounded disabled:opacity-50 disabled:cursor-not-allowed ${
-                    isRectangleSelection ? 'bg-blue-600 hover:bg-blue-500 text-white' : 'bg-gray-600 hover:bg-gray-500 text-white'
-                  }`}
-                  disabled={isScanInProgress}
-                >
-                  <i className="fas fa-vector-square mr-1"></i>
-                  {isRectangleSelection ? 'Stop Selection' : 'Select Area'}
-                </button>
-                <button
-                  onClick={async () => {
-                    if (!microscopeControlService) return;
-                    
-                    if (isScanInProgress) {
-                      // Stop scan logic
-                      try {
-                        if (appendLog) appendLog('Stopping scan...');
-                        
-                        const result = await microscopeControlService.stop_scan_and_stitching();
-                        
-                        if (result.success) {
-                          if (showNotification) showNotification('Scan stop requested', 'success');
-                          if (appendLog) appendLog('Scan stop requested - scan will be interrupted');
-                          setIsScanInProgress(false);
-                          if (setMicroscopeBusy) setMicroscopeBusy(false);
-                          if (setCurrentOperation) setCurrentOperation(null); // Re-enable sidebar
-                        } else {
-                          if (showNotification) showNotification(`Failed to stop scan: ${result.message}`, 'error');
-                          if (appendLog) appendLog(`Failed to stop scan: ${result.message}`);
-                        }
-                      } catch (error) {
-                        if (showNotification) showNotification(`Error stopping scan: ${error.message}`, 'error');
-                        if (appendLog) appendLog(`Error stopping scan: ${error.message}`);
-                      }
-                      return;
-                    }
-                    
-
-                    
-                    // Check if WebRTC is active and stop it to prevent camera resource conflict
-                    const wasWebRtcActive = isWebRtcActive;
-                    if (wasWebRtcActive) {
-                      if (appendLog) appendLog('Stopping WebRTC stream to prevent camera resource conflict during scanning...');
-                      try {
-                        if (toggleWebRtcStream) {
-                          toggleWebRtcStream(); // This will stop the WebRTC stream
-                          // Wait a moment for the stream to fully stop
-                          await new Promise(resolve => setTimeout(resolve, 500));
-                        } else {
-                          if (appendLog) appendLog('Warning: toggleWebRtcStream function not available, proceeding with scan...');
-                        }
-                      } catch (webRtcError) {
-                        if (appendLog) appendLog(`Warning: Failed to stop WebRTC stream: ${webRtcError.message}. Proceeding with scan...`);
-                      }
-                    }
-                    
-                    setIsScanInProgress(true);
-                    if (setMicroscopeBusy) setMicroscopeBusy(true); // Also set global busy state
-                    if (setCurrentOperation) setCurrentOperation('scanning'); // Disable sidebar during scanning
-                    
-                    // Disable rectangle selection during scanning to allow map browsing
-                    setIsRectangleSelection(false);
-                    setRectangleStart(null);
-                    setRectangleEnd(null);
-                    
-                    try {
-                      
-                      
-                      if (appendLog) {
-                        const channelNames = scanParameters.illumination_settings.map(s => s.channel).join(', ');
-                        appendLog(`Starting scan: ${scanParameters.Nx}×${scanParameters.Ny} positions from (${scanParameters.start_x_mm.toFixed(1)}, ${scanParameters.start_y_mm.toFixed(1)}) mm`);
-                        appendLog(`Channels: ${channelNames}`);
-                      }
-                      
-                      const result = await microscopeControlService.normal_scan_with_stitching(
-                        scanParameters.start_x_mm,
-                        scanParameters.start_y_mm,
-                        scanParameters.Nx,
-                        scanParameters.Ny,
-                        scanParameters.dx_mm,
-                        scanParameters.dy_mm,
-                        scanParameters.illumination_settings,
-                        scanParameters.do_contrast_autofocus,
-                        scanParameters.do_reflection_af,
-                        'scan_' + Date.now(),
-                        0, // timepoint index
-                        activeExperiment, // experiment_name parameter
-                        selectedWells, // <-- now supports multi-well
-                        wellPlateType, // wellplate_type parameter
-                        wellPaddingMm, // well_padding_mm parameter
-                        scanParameters.uploading // uploading parameter
-                      );
-                      
-                      if (result.success) {
-                        if (showNotification) showNotification('Scan completed successfully', 'success');
-                        if (appendLog) {
-                          appendLog('Scan completed successfully');
-                          if (wasWebRtcActive) {
-                            appendLog('Note: WebRTC stream was stopped for scanning. Click "Start Live" to resume video stream if needed.');
-                          }
-                        }
-                        setShowScanConfig(false);
-                        setIsRectangleSelection(false);
-                        setRectangleStart(null);
-                        setRectangleEnd(null);
-                        // Enable scan results layer if not already
-                        setVisibleLayers(prev => ({ ...prev, scanResults: true }));
-                        
-                        // Refresh scan results display once after completion
-                        setTimeout(() => {
-                          refreshScanResults();
-                        }, 1000); // Wait 1 second then refresh
-                      } else {
-                        if (showNotification) showNotification(`Scan failed: ${result.message}`, 'error');
-                        if (appendLog) appendLog(`Scan failed: ${result.message}`);
-                      }
-                    } catch (error) {
-                      if (showNotification) showNotification(`Scan error: ${error.message}`, 'error');
-                      if (appendLog) appendLog(`Scan error: ${error.message}`);
-                    } finally {
-                      setIsScanInProgress(false);
-                      if (setMicroscopeBusy) setMicroscopeBusy(false); // Clear global busy state
-                      if (setCurrentOperation) setCurrentOperation(null); // Re-enable sidebar
-                    }
-                  }}
-                className={`px-3 py-1 text-xs text-white rounded disabled:opacity-50 disabled:cursor-not-allowed flex items-center ${
-                  isScanInProgress ? 'bg-red-600 hover:bg-red-500' : 'bg-green-600 hover:bg-green-500'
-                }`}
-                disabled={!microscopeControlService}
-              >
-                {isScanInProgress ? (
-                  <>
-                    <i className="fas fa-stop mr-1"></i>
-                    Stop Scan
-                  </>
-                ) : (
-                  <>
-                    <i className="fas fa-play mr-1"></i>
-                    Start Scan
-                  </>
-                )}
-              </button>
-              
-
-            </div>
-          </div>
-        </div>
-      )}
       {/* Browse Data Modal */}
       {showBrowseDataModal && (
         <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
